fix: remove nested model snapshots from BaseChatSchemaV1 to eliminate entity name collision
The nested @Model classes in BaseChatSchemaV1 duplicated the live top-level
@Model types, producing two Swift types with the same SwiftData entity name
in one binary — causing SIGABRT in CI. Now V1 references the live types
directly; frozen snapshots should be introduced only when V2 is created.

Also applies: versionedSchema API in ModelContainerFactory, ObjectIdentifier-
based assertions in SchemaMigrationTests, deprecation annotation on
BaseChatSchema, and restores model classes from typealiases to standalone
@Model definitions.
